% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_for_mortality.R
\name{clean_data_aggregate}
\alias{clean_data_aggregate}
\title{Function to clean mortality data}
\usage{
clean_data_aggregate(data, information)
}
\arguments{
\item{mortality_data}{Aggregate mortality data}
}
\value{
[data.frame] Clean mortality data
}
\description{
Function to clean mortality data
}
