% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_for_mortality.R
\name{clean_individual_data}
\alias{clean_individual_data}
\title{Function to clean mortality when individual data}
\usage{
clean_individual_data(data, information)
}
\arguments{
\item{data}{Mortality data}
}
\value{
[data.frame] Rotate mortality data
}
\description{
Function to clean mortality when individual data
}
