% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_infos_data.R
\name{f_add_clean_infos}
\alias{f_add_clean_infos}
\title{Function to select the information in the logfile as country name etc..}
\usage{
f_add_clean_infos(item, log_file)
}
\arguments{
\item{item}{name of the folder linked to the log_file}

\item{log_file}{log_file file}
}
\value{
[data.frame] Clean cluster data
}
\description{
Function to select the information in the logfile as country name etc..
}
