% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plausibility_score.R
\name{f_calcul_and_report_plausibility_score}
\alias{f_calcul_and_report_plausibility_score}
\title{Function to generate report containing plausibility score}
\usage{
f_calcul_and_report_plausibility_score(item, information)
}
\arguments{
\item{path_data}{Path for cluster data}
}
\value{
[data.frame] Clean cluster data
}
\description{
Function to generate report containing plausibility score
}
