% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plausibility_score.R
\name{f_calcul_quality_score}
\alias{f_calcul_quality_score}
\title{Function to calculate plausibility score}
\usage{
f_calcul_quality_score(information, nutrition_data, nutrition_data_clean)
}
\arguments{
\item{path_data}{Path for cluster data}
}
\value{
[data.frame] Clean cluster data
}
\description{
Function to calculate plausibility score
}
