% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plausibility_score.R
\name{f_calculate_final_overall_score}
\alias{f_calculate_final_overall_score}
\title{Function to generate the final overall score}
\usage{
f_calculate_final_overall_score(
  flagged_data,
  overall_sex,
  overall_age,
  dps_weight,
  dps_height,
  dps_muac,
  std_weight,
  skew,
  kur,
  poisson_score
)
}
\arguments{
\item{flagged_data}{integer number of flagged data}

\item{overall_sex}{integer number of overall sex}

\item{overall_age}{integer number of overall age}

\item{dps_weight}{integer number of dps weight}

\item{dps_height}{integer number of dps height}

\item{dps_muac}{integer number of dps height}

\item{std_weight}{integer number of std weight}

\item{skew}{integer number of skew}

\item{kur}{integer number of kurotsis}

\item{poisson_score}{integer number of poisson score}
}
\value{
[data.frame] Clean cluster data
}
\description{
Function to generate the final overall score
}
