% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mortality_metadata.R
\name{f_calculate_mortality_metadata}
\alias{f_calculate_mortality_metadata}
\title{Function to update metadata with mortality metadata}
\usage{
f_calculate_mortality_metadata(item, information)
}
\arguments{
\item{item}{name of the folder linked to the log_file}

\item{information}{metadatada of the file selected}
}
\value{
[data.frame] Clean metadda with mortality ones
}
\description{
Function to update metadata with mortality metadata
}
