% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_for_clusters.R
\name{f_clean_and_export_cluster_data}
\alias{f_clean_and_export_cluster_data}
\title{Function to read and save the different cluster data}
\usage{
f_clean_and_export_cluster_data(folder_name, information)
}
\arguments{
\item{folder_name}{Folder name}
}
\value{
[data.frame] Save cluster data in the right folder
}
\description{
Function to read and save the different cluster data
}
