% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_for_nutrition.R
\name{f_clean_and_export_nutritition_data}
\alias{f_clean_and_export_nutritition_data}
\title{Function to read and save the different nutrition data}
\usage{
f_clean_and_export_nutritition_data(
  folder_name,
  information,
  cleaning_criteria
)
}
\arguments{
\item{folder_name}{Folder name}
}
\value{
[data.frame] Save nutrition data in the right folder
}
\description{
Function to read and save the different nutrition data
}
