% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_for_clusters.R
\name{f_clean_cluster_data}
\alias{f_clean_cluster_data}
\title{Function to clean clusters data}
\usage{
f_clean_cluster_data(path_data)
}
\arguments{
\item{path_data}{Path for cluster data}
}
\value{
[data.frame] Clean cluster data
}
\description{
Function to clean clusters data
}
