% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_for_mortality.R
\name{f_clean_mortality_data}
\alias{f_clean_mortality_data}
\title{Function to clean mortality data}
\usage{
f_clean_mortality_data(path_data, information)
}
\arguments{
\item{path_data}{Path for morality data}
}
\value{
[data.frame] Clean mortality data
}
\description{
Function to clean mortality data
}
