% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_for_nutrition.R
\name{f_clean_nutrition_data}
\alias{f_clean_nutrition_data}
\title{Function to clean nutrition data: add flag in the nutrition data + 
generate dataframe with clean nutrition data - without the one flagged +
modify the information by calculating the flag coefficient}
\usage{
f_clean_nutrition_data(path_data, information, cleaning_criteria)
}
\arguments{
\item{path_data}{Path for nutrition data}
}
\value{
[list] Clean nutrition data + information update
}
\description{
Function to clean nutrition data: add flag in the nutrition data + 
generate dataframe with clean nutrition data - without the one flagged +
modify the information by calculating the flag coefficient
}
