% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_all_data.R
\name{f_cleaning}
\alias{f_cleaning}
\title{Function to clean the different folders}
\usage{
f_cleaning(all_survey_folder, cleaning_criteria = "who_2006")
}
\arguments{
\item{folder_name}{folder containing all the data}
}
\value{
[data.frame] Clean cluster data
}
\description{
Function to clean the different folders
}
