% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_for_nutrition.R
\name{f_dico_cleaning_criteria}
\alias{f_dico_cleaning_criteria}
\title{Function to generaye dico cleaning criteria as developed in Marko paper}
\usage{
f_dico_cleaning_criteria()
}
\value{
[dict] List of list containing the cleaning criteria
}
\description{
Function to generaye dico cleaning criteria as developed in Marko paper
}
