% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plausibility_score.R
\name{f_generate_score_table_word}
\alias{f_generate_score_table_word}
\title{Function to modify the style of the overall score list}
\usage{
f_generate_score_table_word(score_list)
}
\arguments{
\item{score_list}{List of scores}
}
\value{
[data.frame] Score list in the right dataframe
}
\description{
Function to modify the style of the overall score list
}
