% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plausibility_score.R
\name{f_modify_table_style}
\alias{f_modify_table_style}
\title{Function to chage the style of table in word document}
\usage{
f_modify_table_style(doc, table)
}
\arguments{
\item{doc}{Word document}

\item{table}{Dataframe to change}
}
\value{
[office doc] return word doc
}
\description{
Function to chage the style of table in word document
}
