% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_plausibility_score.R
\name{f_type_score}
\alias{f_type_score}
\title{Function to qulify the type of overall score}
\usage{
f_type_score(overall_score)
}
\arguments{
\item{overall_score}{integer of the overall score}
}
\value{
[character] type of overall score
}
\description{
Function to qulify the type of overall score
}
