% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mortality_metadata.R
\name{f_update_info_mortality}
\alias{f_update_info_mortality}
\title{Function to update metadata when mortality data is not empty}
\usage{
f_update_info_mortality(mortality_data, information)
}
\arguments{
\item{information}{metadatada of the file selected}
}
\value{
[data.frame] Clean metadda with mortality ones
}
\description{
Function to update metadata when mortality data is not empty
}
