% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mortality_metadata.R
\name{f_update_info_no_mortality_data}
\alias{f_update_info_no_mortality_data}
\title{Function to update metadata when mortality data is empty}
\usage{
f_update_info_no_mortality_data(information)
}
\arguments{
\item{information}{metadatada of the file selected}
}
\value{
[data.frame] Clean metadda with mortality ones
}
\description{
Function to update metadata when mortality data is empty
}
