% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_all_data.R
\name{generate_metadata_file}
\alias{generate_metadata_file}
\title{Function to generate metadata file}
\usage{
generate_metadata_file(all_survey_folder)
}
\value{
[data.frame] metadata
}
\description{
Function to generate metadata file
}
