% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_word_doc.R
\name{generate_word_doc}
\alias{generate_word_doc}
\title{Function to generate word document with plausibility score}
\usage{
generate_word_doc(item, information, score_list, flagged_data)
}
\arguments{
\item{path_data}{Path for cluster data}
}
\value{
[data.frame] Clean cluster data
}
\description{
Function to generate word document with plausibility score
}
