% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_reference_doc}
\alias{get_reference_doc}
\title{Get the path to the reference document}
\usage{
get_reference_doc()
}
\value{
A character string with the path to the reference_doc.docx
}
\description{
Get the path to the reference document
}
