% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_age_dist.R
\name{plot_age_distr}
\alias{plot_age_distr}
\title{Function to plot age distribution}
\usage{
plot_age_distr(nutrition_file, group_by_TEAM = FALSE)
}
\arguments{
\item{nutrition_file}{path of the nutrition file}
}
\value{
plot
}
\description{
Function to plot age distribution
}
