% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_anthro.R
\name{plot_anthro}
\alias{plot_anthro}
\title{Function to plot anthro results could be gaussian curve or probability plot or distribution}
\usage{
plot_anthro(folder_name, exclusion, type_of_plot, group = "All")
}
\arguments{
\item{type_of_plot}{could be gaussian_curve_WHZ, gaussian_curve_WAZ, gaussian_curve_HAZ,
or cum_muac, or prob_plot_WHZ, or prob_plot_HAZ, or prob_plot_WAZ, or
distr_WHZ_2 or distr_WHZ_3 
NOT IMPLEMENTED FOR THE MOMENT distr_GAM or distr_SAM}
}
\value{
plot
}
\description{
Function to plot anthro results could be gaussian curve or probability plot or distribution
}
