% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cumulative_distribution.R
\name{plot_cumulative_curve}
\alias{plot_cumulative_curve}
\title{Function to plot cumulative distribution}
\usage{
plot_cumulative_curve(nutrition_file)
}
\arguments{
\item{nutrition_file}{path of the nutrition file}
}
\value{
plot
}
\description{
Function to plot cumulative distribution
}
