% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distr_per_cluster.R
\name{plot_distribution_per_cluster}
\alias{plot_distribution_per_cluster}
\title{Function to plot distribution per cluster}
\usage{
plot_distribution_per_cluster(nutrition_file, threshold)
}
\arguments{
\item{nutrition_file}{path of the nutrition file}

\item{threshold}{NA or 2 or 3}

\item{anthro_feature}{WHZ or WAZ, or HAZ}
}
\value{
plot
}
\description{
Function to plot distribution per cluster
}
