% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gaussian_curve.R
\name{plot_gaussian_curve}
\alias{plot_gaussian_curve}
\title{Function to plot gaussian curve for anthro results}
\usage{
plot_gaussian_curve(nutrition_file, anthro_feature, group)
}
\arguments{
\item{nutrition_file}{path of the nutrition file}

\item{anthro_feature}{WHZ or WAZ, or HAZ}

\item{group}{NA or SEX}
}
\value{
plot
}
\description{
Function to plot gaussian curve for anthro results
}
