% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_probability_plot.R
\name{plot_proba_curve}
\alias{plot_proba_curve}
\title{Function to plot probbility plot for anthro results}
\usage{
plot_proba_curve(nutrition_file, anthro_feature, group)
}
\arguments{
\item{nutrition_file}{path of the nutrition file}

\item{anthro_feature}{WHZ or WAZ, or HAZ}

\item{group}{NA or SEX}
}
\value{
plot
}
\description{
Function to plot probbility plot for anthro results
}
