% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_for_mortality.R
\name{rotate_data}
\alias{rotate_data}
\title{Function to rotate mortality when individual data}
\usage{
rotate_data(df)
}
\arguments{
\item{df}{Mortality data}
}
\value{
[data.frame] Rotate mortality data
}
\description{
Function to rotate mortality when individual data
}
