% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_per_team.R
\name{table_analysis_per_team}
\alias{table_analysis_per_team}
\title{Function to generate table analysis per team}
\usage{
table_analysis_per_team(folder_name)
}
\arguments{
\item{folder_name}{name of the folder where the data are available}
}
\value{
plot
}
\description{
Function to generate table analysis per team
}
